pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cambrian"
version = "1.2.1"
description = "A Fully Open, Vision-Centric Exploration of Multimodal LLMs."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "torch==2.2.0", "torchvision==0.17.0",
    "transformers==4.37.0", "tokenizers==0.15.0", "sentencepiece==0.1.99", "shortuuid",
    "accelerate==0.23.0", "peft==0.4.0",
    "pydantic<2,>=1", "markdown2[all]", "numpy", "scikit-learn==1.2.2",
    "gradio==3.35.2", "gradio_client==0.2.9",
    "requests", "httpx==0.24.0", "uvicorn", "fastapi",
    "einops==0.6.1", "einops-exts==0.0.4", "timm==0.9.16",
    "open_clip_torch", "diffusers[torch]", "torchtext==0.17.0",
    "ezcolorlog", "gcsfs"
]

[project.optional-dependencies]
train = ["deepspeed==0.9.5", "ninja", "wandb"]
gpu = ["bitsandbytes==0.41.0", "deepspeed==0.9.5", "ninja", "wandb"]
tpu = ["ninja", "wandb"]

[project.urls]
"Homepage" = "https://cambrian-mllm.github.io/"

[tool.setuptools.packages.find]
exclude = ["assets*", "benchmark*", "docs", "dist*", "playground*", "scripts*", "tests*"]

[tool.wheel]
exclude = ["assets*", "benchmark*", "docs", "dist*", "playground*", "scripts*", "tests*"]
